[#1590] Fix api host variable injection in chatplugin
