+++ frontend/chat-plugin/index.ts
@@ -1,1 +1,2 @@
 export * from './src/AiryChatPlugin';
+export * from './src/config';
