+++ frontend/chat-plugin/src/api/index.tsx
@@ -1,17 +1,13 @@
 import {QuickReplyCommand, SuggestionResponse, TextContent} from 'render/providers/chatplugin/chatPluginModel';
 import {setResumeTokenInStorage} from '../storage';
 
-declare const window: {
-  airy: {
-    host: string;
-    channelId: string;
-  };
+let host;
+export const setApiHost = apiHost => {
+  host = apiHost;
 };
 
-const API_HOST = window.airy ? window.airy.host : process.env.API_HOST;
-
 export const sendMessage = (message: TextContent | SuggestionResponse | QuickReplyCommand, token: string) => {
-  return fetch(`${API_HOST}/chatplugin.send`, {
+  return fetch(`${host}/chatplugin.send`, {
     method: 'POST',
     body: JSON.stringify(convertToBody(message)),
     headers: {
@@ -39,7 +35,7 @@ const convertToBody = (message: TextContent | SuggestionResponse | QuickReplyCom
 };
 
 export const getResumeToken = async (channelId: string, authToken: string) => {
-  const resumeChat = await fetch(`${API_HOST}/chatplugin.resumeToken`, {
+  const resumeChat = await fetch(`${host}/chatplugin.resumeToken`, {
     method: 'POST',
     body: JSON.stringify({}),
     headers: {
@@ -53,7 +49,7 @@ export const getResumeToken = async (channelId: string, authToken: string) => {
 
 export const start = async (channelId: string, resumeToken: string) => {
   try {
-    const response = await fetch(`${API_HOST}/chatplugin.authenticate`, {
+    const response = await fetch(`${host}/chatplugin.authenticate`, {
       method: 'POST',
       body: JSON.stringify({
         channel_id: channelId,
