+++ frontend/chat-plugin/src/config.ts
@@ -1,18 +1,29 @@
-import {Config} from './App';
-
 export type RenderCtrl = {
   toggleHideChat: () => void;
 };
 
 export type RenderProp = (ctrl?: RenderCtrl) => JSX.Element;
 
+export type Config = {
+  welcomeMessage?: {};
+  headerText?: string;
+  headerTextColor?: string;
+  backgroundColor?: string;
+  primaryColor?: string;
+  accentColor?: string;
+  bubbleIcon?: string;
+  sendMessageIcon?: string;
+  showMode: boolean;
+};
+
 export type AuthConfiguration = {
   channelId: string;
   resumeToken?: string;
-  config?: Config;
 };
 
 export type AiryChatPluginConfiguration = AuthConfiguration & {
+  apiHost: string;
+  config?: Config;
   headerBarProp?: RenderProp;
   inputBarProp?: RenderProp;
   airyMessageProp?: RenderProp;
